fix(cloudflare)!: hand Durable Object stubs Skyzen requests and responses
`CfDurableObjectStub::fetch` already returned a `Send` future, but it returned a
`web_sys::Response` — and a handler holds that response across the await that
reads its body, which made the handler itself `!Send`. So every route touching a
Durable Object had to wrap its own body in `worker::send::SendFuture::new(async
move { ... }).await`, asserting a property the caller cannot violate anyway, and
forgetting it produced a `Future is not Send` error pointing at the handler
rather than at the stub.

`fetch` now takes a `skyzen::Request` and returns a `skyzen::Response` — the same
signature `NativeDurableObjectStub` has natively, so one handler body works
against a simulated object in a test and a real one on Workers. Nothing in the
returned value is a JS handle, so the ceremony is gone.

That needed one conversion between WinterCG request/response objects and
Skyzen's own, in both directions. There were already two hand-rolled copies of
half of it — `src/runtime/wasm.rs` for the Worker entry point and
`cloudflare/src/durable/glue.rs` for the Durable Object glue — which had already
drifted: one appended repeated headers while the other overwrote them, one
streamed bodies while the other buffered them. Both now go through
`skyzen::runtime::wasm::{from_js_request, into_js_request, from_js_response,
into_js_response}`, which stream in both directions and append. `CfAssets::respond`
loses its third partial copy.

A `101` keeps its socket across the boundary: an incoming one has it moved into
the response extensions, and the outgoing conversion builds the JS response from
the whole `Response` — status, headers and socket — so proxying a websocket
upgrade to a Durable Object is now just returning what the stub handed back.

The compile-time `assert_send` block gains the shape that actually regressed: a
whole handler that awaits the stub, then awaits the body read, then dispatches
again.

Closes #17
